Refactor Vines functionality: rename VinesHelper to VinesIndexHelper, add new VinesHelper for WordStudyPanel, and implement index generation and rendering
