+++ app/frontend/components/tool_panel/word_study_panel/word_study_panel.js
@@ -22,6 +22,7 @@ const RobinsonGreekMorphologyParser = require('./robinson_greek_morphology_parse
 const PackardGreekMorphologyParser = require('./packard_greek_morphology_parser.js');
 const OpenScripturesHebrewMorphologyParser = require('./oshm_morphology_parser.js');
 const StrongsOccurrencesHelper = require('./strongs_occurrences_helper.js');
+const VinesHelper = require('./vines_helper.js');
 
 let jsStrongs = null;
 
@@ -51,6 +52,7 @@ class WordStudyPanel {
     this._packardMorphologyParser = new PackardGreekMorphologyParser();
     this._oshmMorphologyParser = new OpenScripturesHebrewMorphologyParser();
     this._occurrencesHelper = new StrongsOccurrencesHelper(this);
+    this._vinesHelper = new VinesHelper(this);
 
     this.wordStudyPanelCopyButton.on('click', (event) => {
       event.preventDefault();
@@ -176,7 +178,7 @@ class WordStudyPanel {
     this.wordStudyPanelCopyButton.show();
 
     this._occurrencesHelper.attachOccurrencesEventHandlers();
-    this.attachVinesEventHandlers();
+    this._vinesHelper.attachVinesEventHandlers();
   }
 
   handleCopyButtonClick() {
@@ -431,7 +433,7 @@ class WordStudyPanel {
     }
 
     let extraDictContent = await this.getExtraDictionaryContent(lang, strongsEntry);
-    let vinesContent = await this.getVinesContent(strongsEntry);
+    let vinesContent = await this._vinesHelper.getVinesContent(strongsEntry);
     let relatedStrongsContent = await this.getRelatedStrongsContent(strongsEntry.references);
     
     const moduleInfoButtonTitle = i18n.t('menu.show-module-info');
@@ -628,129 +630,6 @@ class WordStudyPanel {
         <i class='fas fa-copy copy-icon'></i>
       </div>
     `;
-  }
-
-  async getVinesContent(strongsEntry) {
-    // Vines only applies to Greek Strong's numbers
-    if (!strongsEntry.key || strongsEntry.key[0] !== 'G') {
-      return '';
-    }
-
-    // Check if the Vines module is installed
-    var dictModules = await ipcNsi.getAllLocalModules('DICT');
-    var vinesInstalled = dictModules.some((m) => m.name === 'Vines');
-
-    if (!vinesInstalled) {
-      return this._getVinesNotInstalledHtml();
-    }
-
-    var indexExists = await ipcGeneral.vinesIndexExists();
-    if (!indexExists) {
-      return this._getVinesIndexNotAvailableHtml();
-    }
-
-    return await this._getVinesLinksHtml(strongsEntry.key);
-  }
-
-  _getVinesNotInstalledHtml() {
-    return `
-      <div id='vines-box'>
-        <div class='dictionary-section'>
-          <div class='bold word-study-title' style='margin-bottom: 0.5em'>Vine's Expository Dictionary</div>
-          <div class='dictionary-content'>
-            <p>${i18n.t('word-study-panel.vines-not-installed')}</p>
-          </div>
-        </div>
-        <hr></hr>
-      </div>
-    `;
-  }
-
-  _getVinesIndexNotAvailableHtml() {
-    return `
-      <div id='vines-box'>
-        <div class='dictionary-section'>
-          <div class='bold word-study-title' style='margin-bottom: 0.5em'>Vine's Expository Dictionary</div>
-          <div class='dictionary-content'>
-            <p>${i18n.t('word-study-panel.vines-index-not-available')}</p>
-            <button id='generate-vines-index-button'
-                    class='fg-button ui-corner-all ui-state-default'>
-              ${i18n.t('word-study-panel.generate-index')}
-            </button>
-          </div>
-        </div>
-        <hr></hr>
-      </div>
-    `;
-  }
-
-  async _getVinesLinksHtml(strongsKey) {
-    var vinesKeys = await ipcGeneral.getVinesKeysForStrongs(strongsKey);
-
-    if (!vinesKeys || vinesKeys.length === 0) {
-      return '';
-    }
-
-    let links = vinesKeys.map((key) => {
-      return `<a href="sword://Vines/${key}" style="cursor: pointer;">${key}</a>`;
-    }).join('<br>');
-
-    return `
-      <div id='vines-box'>
-        <div class='dictionary-section'>
-          <div class='bold word-study-title' style='margin-bottom: 0.5em'>Vine's Expository Dictionary</div>
-          <div class='dictionary-content'>${links}</div>
-        </div>
-        <hr></hr>
-      </div>
-    `;
-  }
-
-  attachVinesEventHandlers() {
-    const panelContent = this.wordStudyPanelContent[0];
-
-    let generateButton = panelContent.querySelector('#generate-vines-index-button');
-    if (generateButton != null) {
-      generateButton.addEventListener('click', () => {
-        this.handleGenerateVinesIndex(generateButton);
-      });
-    }
-  }
-
-  async handleGenerateVinesIndex(button) {
-    const vinesBox = document.getElementById('vines-box');
-
-    if (vinesBox == null) {
-      return;
-    }
-
-    vinesBox.innerHTML = `
-      <div class='dictionary-section'>
-        <div class='bold word-study-title' style='margin-bottom: 0.5em'>Vine's Expository Dictionary</div>
-        <div class='dictionary-content'>
-          <div id='vines-index-progress-bar' class='progress-bar'>
-            <div class='progress-label'>0%</div>
-          </div>
-        </div>
-      </div>
-      <hr></hr>
-    `;
-
-    const $progressBar = $('#vines-index-progress-bar');
-    uiHelper.initProgressBar($progressBar);
-
-    await ipcGeneral.buildVinesIndex((progress) => {
-      $progressBar.progressbar("value", progress.totalPercent);
-    });
-
-    const currentStrongsEntry = this.currentStrongsEntry;
-    if (currentStrongsEntry != null) {
-      vinesBox.outerHTML = await this._getVinesLinksHtml(currentStrongsEntry.key);
-
-      // Re-initialize sword:// links for the new content
-      const swordUrlHelper = require('../../../helpers/sword_url_helper.js');
-      swordUrlHelper.initSwordUrlLinks(this.wordStudyPanelContent[0], null);
-    }
   }
 
   async getDictionaryEntry(moduleCode, strongsEntry) {
